feat(Information.java): destination 및 README.md 자료형 변경
- destination 자료형 long -> int 변경
- README.md 내부 long -> int 변경

+++ src/main/java/racingcar/Information.java
@@ -2,7 +2,7 @@ package racingcar;
 
 public class Information {
     String name;
-    long destination;
+    int destination;
 
     Information(String name) {
         this.name = name;
